requirements.txt
# MCP Server Dependencies
mcp>=1.2.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For enhanced testing
responses>=0.23.0